实现了接单平台只显示始发点距离司机小于10公里的订单的逻辑

+++ src/cn/edu/zjut/dao/OrderDAO.java
@@ -82,8 +82,8 @@ public class OrderDAO extends BaseHibernateDAO implements IOrderDAO {
 			List<Order> orders=findorders("from Order as user where passengerID="+i);
 			for(int j=0;j<orders.size();j++) {
 			Order example=orders.get(j);
-			 if(example.getIsEstimatedP()==false) {
-			session.delete(example);
+
+	 		if(example.getIsEstimatedP()==false) {			session.delete(example);
 			 }
 			}
 			tran.commit();
@@ -97,8 +97,8 @@ public class OrderDAO extends BaseHibernateDAO implements IOrderDAO {
 			List<Order> orders=findorders("from Order as user where passengerID="+i);
 			for(int j=0;j<orders.size();j++) {
 			Order example=orders.get(j);
-			 if(example.getIsEstimatedP()==false) {
-			session.delete(example);
+			
+ 			if(example.getIsEstimatedP()==false) {			session.delete(example);
 			 }
 			}
 			int k=order.getDriver().getDriverID();
@@ -145,8 +145,7 @@ public class OrderDAO extends BaseHibernateDAO implements IOrderDAO {
         order.setSum(order.getSum());
 		order.setEstimateptod(score1);
 		order.setIsCompleted(2);
-		order.setIsEstimatedP(true);
-		session.update(order);
+		order.setIsEstimatedP(true);		session.update(order);
 		tran.commit();
 	}
 }
